ДЗ по теме "Работа с файлами". Все методы работают, только если задавать в ручную id для таблицы report и reportWithPath

+++ src/main/java/ru/skypro/lessons/springboot/weblibrary_1/DTO/EmployeeReport.java
@@ -11,8 +11,8 @@ import java.io.Serializable;
 @Setter
 public class EmployeeReport implements Serializable {
     private int department;
-    private int countOfEmployees;
+    private Long countOfEmployees;
     private int minSalary;
     private int maxSalary;
-    private int averageSalary;
+    private double averageSalary;
 }
